print list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -26,7 +26,7 @@ int _strlen(char *str)
  */
 size_t print_list(const list_t *h)
 {
-	size_t counter = 0;
+	unsigned int counter = 0;
 
 	while (h)
 	{
